Add solution for problem `D - Logical Expression` Link: https://atcoder.jp/contests/abc189/tasks/abc189_d Tags: AtCoder, *400 Ref: https://atcoder.jp/contests/abc189/submissions/19678459

+++ atcoder/abc189/D.cpp
@@ -143,7 +143,7 @@ void solve(){
 	rep(i,n){
 		string x;
 		cin>>x;
-		if(x=="OR") ans+=(1LL<<(i+1));
+		if(x=="OR") ans+=(1LL<<(ll)(i+1));
 	}
 	cout<<(ans+1ll)<<endl;
 }
